+++ DOCS/CHANGELOG.md
@@ -52,6 +52,7 @@
 - new macros:
   - `define_panic_handler!`, `define_static_map!`, `methods_as_fns!`.
   - `js_reexport`, `maybe!`, `miniquad!`, `strjoin!`, `xorshift_custom!`.
+  - `os_print`, `os_println`, `os_eprint`, `os_eprintln`.
 - new modules:
   - data: `{codec::{self, crypto, radix}, list, key, table, uid}`.
   - lang: `{dsl, ffi::{self, c, glsl}, i18n, ling::{art, nat}}`.